modify_source_node
