+++ source_node/include/navigation_source.h
@@ -1,16 +1,12 @@
-/** brief:头文件声明了组合导航资源节点，继承source_node
- *  note:
- *  
- */
-
-#ifndef NAVIGATION_SOURCE_H
-#define NAVIGATION_SOURCE_H
+#pragma once
 
 #include "source_node.h"
 #include "gps.h"
 #include "imu.h"
+#include "ros/ros.h"
+#include "robot_msgs/Cmd.h"
 
-class NavigationSource : public SourceNode{
+class NavigationSource{
 public:
     NavigationSource();
     ~NavigationSource(){};
@@ -24,8 +20,15 @@ public:
     void CmdCallback(const robot_msgs::CmdConstPtr &msg);
     GPS gps_;
     IMU imu_;
+
+    ros::Publisher  state_pub;
+    ros::Subscriber cmd_sub;
+
+    State node_state;
+    Error node_error;
+    uint8_t update_frequence;
+    std::string node_name;
 };
 
 
 
-#endif
