+++ source_node/include/source_node.h
@@ -1,14 +1,5 @@
-/** brief:头文件声明了source节点的基类，SourceNode
- *  note:
- *  
- */
+#pragma once
 
-#ifndef SOURCE_NODE_H
-#define SOURCE_NODE_H
-#include <thread>
-#include <string>
-#include "robot_msgs/SourceNodeMsg.h"
-#include "robot_msgs/Cmd.h"
 #include "ros/ros.h"
 
 #define pi 3.14
@@ -38,27 +29,3 @@ enum class Error{
     CONFIG_FAILED = 1
 };
 
-
-class SourceNode{
-public:
-    SourceNode()=default;
-    ~SourceNode()=default;
-    virtual bool  LoadConfig(std::string file){}
-    //这些函数应该在cmd的callback中被调用
-    virtual State Start(){}
-    virtual State Stop(){}
-    virtual State Exit(){}
-    virtual State Pause(){}
-    virtual State Resume(){}
-    virtual void  UpdateState(){}
-    virtual void CmdCallback(const robot_msgs::CmdConstPtr &msg){}
-    ros::Publisher  state_pub;
-    ros::Subscriber cmd_sub;
-    State node_state;
-    Error node_error;
-    uint8_t update_frequence;
-    std::string node_name;
-};
-
-
-#endif
