--- source_node/src/source_node.cpp
@@ -1,2 +1,0 @@
-
-
